Nave 7 pruebas

+++ ControlPicking/ControlPicking/Models/Listas.cs
@@ -8,12 +8,12 @@
         public int Orden { get; set; }
         public int Line { get; set; }
 
-        public int IdPick { get; set; }
+        public long IdPick { get; set; }
 
     }
     public class Detalle
     {
-        public int IdPick { get; set; }
+        public long IdPick { get; set; }
         public string Pick_Lnp { get; set; }
         public string Item { get; set; }
         public int Qty { get; set; }
